refactor(session): establish AppSession Flow

+++ ball/profile/interaction/src/commonMain/kotlin/kinetickk/ball/profile/interaction/lab/api/LabFeature.kt
@@ -31,7 +31,7 @@ sealed interface LabOutput {
 interface LabFeature {
     @Composable
     fun Content(
-        routeToken: Int,
+        routeToken: Long,
         onOutput: (LabOutput) -> Unit,
     )
 }
